Copy edits to lesson 2 of unit 4
I think the lesson would flow better if the "Why Use APIs?" section comes before the "Web APIs" section, and the "Web APIs" section is followed by the "How an API Fits in With SPAs" section. Also, can we edit the image under the "APIs in the Wild" section to read: "PLANET OF THE APIs" instead of ""PLANET OF THE APIS"?